NFC: Move CombineTransferReadOpBroadcast to Common/GPUPatterns
This is in preparation to be used by SPIR-V CodeGen flow.
