+++ compiler/src/iree/compiler/Codegen/Common/GPUPatterns.cpp
@@ -142,6 +142,40 @@ struct FlattenTransferReadOp : public OpRewritePattern<vector::TransferReadOp> {
   }
 };
 
+// Merges transpose op into the transfer read op. Transpose are not supported on
+// MMA types but MMA load can transpose the matrix when loading.
+struct CombineTransferReadOpBroadcast final
+    : public OpRewritePattern<vector::BroadcastOp> {
+  using OpRewritePattern<vector::BroadcastOp>::OpRewritePattern;
+
+  LogicalResult matchAndRewrite(vector::BroadcastOp op,
+                                PatternRewriter &rewriter) const override {
+    auto transferReadOp =
+        op.getSource().getDefiningOp<vector::TransferReadOp>();
+    if (!transferReadOp || transferReadOp.getMask() ||
+        transferReadOp.hasOutOfBoundsDim()) {
+      return failure();
+    }
+    int64_t rankDiff =
+        op.getVectorType().getRank() - transferReadOp.getVectorType().getRank();
+    SmallVector<AffineExpr> exprs(rankDiff, rewriter.getAffineConstantExpr(0));
+    ArrayRef<AffineExpr> originalExpr =
+        transferReadOp.getPermutationMap().getResults();
+    exprs.append(originalExpr.begin(), originalExpr.end());
+    AffineMap newMap =
+        AffineMap::get(transferReadOp.getPermutationMap().getNumDims(),
+                       transferReadOp.getPermutationMap().getNumSymbols(),
+                       exprs, op.getContext());
+    ArrayAttr inBounds = rewriter.getBoolArrayAttr(
+        SmallVector<bool>(op.getVectorType().getRank(), true));
+    rewriter.replaceOpWithNewOp<vector::TransferReadOp>(
+        op, op.getType(), transferReadOp.getSource(),
+        transferReadOp.getIndices(), newMap, transferReadOp.getPadding(),
+        transferReadOp.getMask(), inBounds);
+    return success();
+  }
+};
+
 template <typename T>
 using LinalgPromotionPattern =
     mlir::iree_compiler::IREE::LinalgExt::LinalgPromotionPattern<T>;
@@ -163,6 +197,11 @@ static LogicalResult contractOpFilter(Operation *op) {
 void populateVectorTransferToGPUMMAPreparationPatterns(
     RewritePatternSet &patterns) {
   patterns.add<FlattenTransferReadOp>(patterns.getContext());
+}
+
+void populateCombineVectorTransferReadBroadcastPatterns(
+    RewritePatternSet &patterns) {
+  patterns.add<CombineTransferReadOpBroadcast>(patterns.getContext());
 }
 
 void populateContractPromotionPatterns(RewritePatternSet &patterns,
